Correct `project` method
The `project` method in `FloatingVector2` was multiplying by the length of the first vector, but it should have been dividing by the length of the second vector.
This change fixes the formula, which now also requires checking to avoid division by 0. If projecting onto a vector shorter than epsilon, a zero vector is returned.